Refactor AFD
