+++ README.md
@@ -9,11 +9,12 @@
 🚀 Welcome to the MUI Auto Form Dialog (AFD) project! The goal of this project is to simplify the creation of form dialogs using Material-UI (MUI) components.
 
 ## Project Overview
-**MUI Auto Form Dialog (AFD)** is a React component that automates the process of building form dialogs through a JSON configuration. Say goodbye to manually coding UI components or wrestling with complex styling – AFD takes care of it all. This component streamlines your workflow, enabling you to focus on defining your form's structure and data requirements.
+**MUI Auto Form Dialog (AFD)** is a React component that automates the process of building form dialogs through a TypeScript-based configuration object configuration. Say goodbye to manually coding UI components or wrestling with complex styling – AFD takes care of it all. This component streamlines your workflow, enabling you to focus on defining your form's structure and data requirements.
 
 ## Key Advantages
-- **Low-Code Approach:** AFD uses a modular JSON configuration to generate clean and polished Form Dialogs every time without exhaustive coding.
-- **Simplified UI Creation:** Seamlessly generates Material-UI components based on your JSON configuration, streamlining the UI assembly process.
+- **Low-Code Approach:** AFD uses a modular configuration object configuration to generate clean and polished Form Dialogs every time without exhaustive coding.
+- **Simplified UI Creation:** Seamlessly generates Material-UI components based on your configuration object configuration, streamlining the UI assembly process.
 - **Automatic CSS Styling:** AFD handles CSS styling automatically, guaranteeing uniformity and responsiveness across your UI design.
 Seamless
-- **API Integration:** AFD uses Axios to send HTTP requests. The query and request type get defined from the values in the JSON configuration.
+- **API Integration:** AFD uses Axios to send HTTP requests. The query and request type get defined from the values in the configuration object configuration.
+- **TypeScript Support:** AFD is written in TypeScript, enabling you to define your form's data requirements and structure with ease.
