+++ src/App.tsx
@@ -2,6 +2,8 @@ import React, { useState } from 'react';
 import { Button, Grid } from '@mui/material';
 import AutoDialog from './components/AutoDialog/AutoDialog';
 
+import { exampleConfiguration } from './exampleConfiguration';
+
 export default function App() {
 
   const [showAutoDialog, setShowAutoDialog] = useState<boolean>(false);
@@ -23,73 +25,7 @@ export default function App() {
           open={showAutoDialog} 
           title={'Auto Dialog'} 
           details={'This Dialog should help you develope faster!!'}
-          data={{
-            action: 'post',
-            query: '',
-            formComponents: [
-              {
-                controlType: 'TextField',
-                size: {
-                  sx: 6,
-                },
-                controlProps: {
-                  id: 'fName',
-                  label: 'First Name',
-                  required: true,
-                },
-              },
-              {
-                controlType: 'TextField',
-                size: {
-                  sx: 6,
-                },
-                controlProps: {
-                  id: 'lName',
-                  label: 'Last Name',
-                  required: true,
-                },
-              },
-              {
-                controlType: 'Autocomplete',
-                size: {
-                  sx: 4,
-                },
-                controlProps: {
-                  id: 'gender',
-                  label: 'Gender',
-                  required: true,
-                  noOptionsText: "No Genders?",
-                  options: ["male","female","other"]
-                  
-                },
-              },
-              {
-                controlType: 'NumberField',
-                size: {
-                  sx: 2,
-                },
-                controlProps: {
-                  id: 'age',
-                  label: 'age',
-                  required: true,
-                  min: 1,
-                  max: 120,
-                  disalowedValues: [69],
-                },
-              },
-              {
-                controlType: 'TextField',
-                size: {
-                  sx: 12,
-                },
-                controlProps: {
-                  id: 'email',
-                  label: 'e-mail',
-                  required: true,
-                },
-              },
-            ]
-          }}
+          data={exampleConfiguration}
           onClose={()=>setShowAutoDialog(false)}
         />
       )}
